+++ ONT3001_Assignment/Controllers/CustomersController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using System.Threading.Tasks;
 using ONT3001_Assignment.Models;
 
 namespace ONT3001_Assignment.Controllers
@@ -15,19 +16,23 @@ namespace ONT3001_Assignment.Controllers
         private StoreDBEntities1 db = new StoreDBEntities1();
 
         // GET: Customers
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.Customers.ToList());
+
+            var customer = await db.Customers.OrderByDescending(c => c.CustomerID).Take(10).ToListAsync(); 
+
+
+            return View(customer);
         }
 
         // GET: Customers/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Customer customer = db.Customers.Find(id);
+            Customer customer = await db.Customers.FindAsync(id);
             if (customer == null)
             {
                 return HttpNotFound();
@@ -46,12 +51,20 @@ namespace ONT3001_Assignment.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "CustomerID,FirstName,LastName,Email,PhoneNumber")] Customer customer)
+        public async Task<ActionResult> Create([Bind(Include = "CustomerID,FirstName,LastName,Email,PhoneNumber")] Customer customer)
         {
+            // Doesnt allow people to have the same 
+            bool emailExists = await db.Customers.AnyAsync(c => c.Email == customer.Email);
+            if (emailExists)
+            {
+                ModelState.AddModelError("Email", "A customer with this email already exists.");
+                return View(customer);
+            }
+
             if (ModelState.IsValid)
             {
                 db.Customers.Add(customer);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -59,14 +72,14 @@ namespace ONT3001_Assignment.Controllers
         }
 
         // GET: Customers/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null) //null ID check
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Customer customer = db.Customers.Find(id);
-            if (customer == null)
+            Customer customer = await db.Customers.FindAsync(id);
+            if (customer == null) //Check if custmer exists
             {
                 return HttpNotFound();
             }
@@ -78,25 +91,33 @@ namespace ONT3001_Assignment.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "CustomerID,FirstName,LastName,Email,PhoneNumber")] Customer customer)
+        public async Task<ActionResult> Edit([Bind(Include = "CustomerID,FirstName,LastName,Email,PhoneNumber")] Customer customer)
         {
+            // Check if they are the same. 
+            bool emailExists = await db.Customers.AnyAsync(c => c.Email == customer.Email
+                                                           && c.CustomerID != customer.CustomerID);
+            if (emailExists)
+            {
+                ModelState.AddModelError("Email", "A customer with this email already exists.");
+                return View(customer);
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(customer).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(customer);
         }
 
         // GET: Customers/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null)  // Null ID check 
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Customer customer = db.Customers.Find(id);
+            Customer customer =await db.Customers.FindAsync(id);
             if (customer == null)
             {
                 return HttpNotFound();
@@ -107,11 +128,20 @@ namespace ONT3001_Assignment.Controllers
         // POST: Customers/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Customer customer = db.Customers.Find(id);
+            Customer customer = await db.Customers.FindAsync(id);
+
+            if (customer == null) return HttpNotFound(); // Check if customer actaully exists 
+
+            if (customer.Orders.Any())
+            {
+                ModelState.AddModelError(String.Empty, "Cannot delete, customer has order in place.");
+                return View(customer); 
+                
+            }
             db.Customers.Remove(customer);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
             return RedirectToAction("Index");
         }
