+++ ONT3001_Assignment/Controllers/ProductsController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using System.Threading.Tasks; 
 using ONT3001_Assignment.Models;
 
 namespace ONT3001_Assignment.Controllers
@@ -15,28 +16,31 @@ namespace ONT3001_Assignment.Controllers
         private StoreDBEntities1 db = new StoreDBEntities1();
 
         // GET: Products
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
             // Use .Where() to only show items in stock, and .OrderBy() to sort by price
-            var availableProducts = db.Products
+            var products = await db.Products
                                       .Where(p => p.StockQuantity > 0)
                                       .OrderBy(p => p.Price)
-                                      .ToList();
+                                      .ToListAsync();
 
-            return View(availableProducts);
+            return View(products);
 
-            var products = db.Products.Include(p => p.Category);
-            return View(products.ToList());
+           
         }
 
         // GET: Products/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Product product = db.Products.Find(id);
+
+            Product product = await db.Products
+                                     .Where(p => p.ProductID == id)
+                                     .FirstOrDefaultAsync();
+
             if (product == null)
             {
                 return HttpNotFound();
@@ -56,12 +60,39 @@ namespace ONT3001_Assignment.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ProductID,CategoryID,ProductName,Price,StockQuantity")] Product product)
+        public async Task<ActionResult>  Create([Bind(Include = "ProductID,CategoryID,ProductName,Price,StockQuantity")] Product product)
         {
+
+            // validating price- price cant be 0 
+            if (product.Price <= 0)
+            {
+                ModelState.AddModelError("Price", "Price cant be zero.");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
+            // the stock cant be below 0 
+            if (product.StockQuantity < 0)
+            {
+                ModelState.AddModelError("StockQuantity", "Cant have negative stock.");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
+            // Catagory needs to avalible to actaully be placed in 
+            bool categoryExists = await db.Categories.AnyAsync(c => c.CategoryID == product.CategoryID);
+
+            if (!categoryExists)
+            {
+                ModelState.AddModelError("CategoryID", "Catagory doesnt excists");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
             if (ModelState.IsValid)
             {
                 db.Products.Add(product);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -70,13 +101,15 @@ namespace ONT3001_Assignment.Controllers
         }
 
         // GET: Products/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Product product = db.Products.Find(id);
+            Product product = await db.Products.Where(p=> p.ProductID == id)
+                                               .FirstOrDefaultAsync();
+            
             if (product == null)
             {
                 return HttpNotFound();
@@ -90,12 +123,38 @@ namespace ONT3001_Assignment.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ProductID,CategoryID,ProductName,Price,StockQuantity")] Product product)
+        public async Task<ActionResult> Edit([Bind(Include = "ProductID,CategoryID,ProductName,Price,StockQuantity")] Product product)
         {
+            // procduct price cant be 0 
+            if (product.Price <= 0)
+            {
+                ModelState.AddModelError("Price", "Price cant be zero.");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
+            // The Stock cant be less than 0 
+            if (product.StockQuantity < 0)
+            {
+                ModelState.AddModelError("StockQuantity", "Cant have negative stock.");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
+            // The category has to exists
+            bool categoryExists = await db.Categories.AnyAsync(c => c.CategoryID == product.CategoryID);
+            if (!categoryExists)
+            {
+                ModelState.AddModelError("CategoryID", "Catagory doesnt exist.");
+                ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
+                return View(product);
+            }
+
+
             if (ModelState.IsValid)
             {
                 db.Entry(product).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             ViewBag.CategoryID = new SelectList(db.Categories, "CategoryID", "CategoryName", product.CategoryID);
@@ -103,13 +162,16 @@ namespace ONT3001_Assignment.Controllers
         }
 
         // GET: Products/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Product product = db.Products.Find(id);
+
+            Product product = await db.Products.Where(p=>p.ProductID == id)
+                                               .FirstOrDefaultAsync(); 
+
             if (product == null)
             {
                 return HttpNotFound();
@@ -120,11 +182,14 @@ namespace ONT3001_Assignment.Controllers
         // POST: Products/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async  Task<ActionResult> DeleteConfirmed(int id)
         {
-            Product product = db.Products.Find(id);
+            Product product = await db.Products.Where(p => p.ProductID == id)
+                                               .FirstOrDefaultAsync();
+            if (product == null) return HttpNotFound();
+
             db.Products.Remove(product);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
             return RedirectToAction("Index");
         }
